move function 'appendQueryString' to 'Utilities'
