+++ lib/index.js
@@ -331,8 +331,10 @@ const appendDataContext = function (plugin, settings) {
 
       // append tags from document request to JSON request
       settings.jsonPath = request.query.tags
-        ? appendQueryString(settings.jsonPath, 'tags', request.query.tags)
-        : appendQueryString(settings.jsonPath, null, null)
+        ? Utilities.appendQueryString(settings.jsonPath, 'tags', request.query.tags)
+        : Utilities.appendQueryString(settings.jsonPath, null, null);
+
+      console.log('result', settings.jsonPath);
 
       const prefixedSettings = Hoek.clone(settings);
       if (routePrefix) {
@@ -355,26 +357,6 @@ const appendDataContext = function (plugin, settings) {
 
     return h.continue;
   });
-};
-
-/**
- * appends a querystring to a url path - will overwrite existing values
- *
- * @param  {string} url
- * @param  {string} qsName
- * @param  {string} qsValue
- * @return {string}
- */
-const appendQueryString = function (url, qsName, qsValue) {
-  const urlObj = Url.parse(url);
-  if (qsName && qsValue) {
-    urlObj.query = Querystring.parse(qsName + '=' + qsValue);
-    urlObj.search = '?' + encodeURIComponent(qsName) + '=' + encodeURIComponent(qsValue);
-  } else {
-    urlObj.search = '';
-  }
-
-  return urlObj.format(urlObj);
 };
 
 /**
