+++ lib/utilities.js
@@ -1,5 +1,7 @@
 const Hoek = require('@hapi/hoek');
 const Joi = require('joi');
+const Url = require('url');
+const Querystring = require('querystring');
 
 const utilities = (module.exports = {});
 
@@ -465,4 +467,24 @@ utilities.assignVendorExtensions = function(target, source) {
   }
 
   return target;
+};
+
+/**
+ * appends a querystring to a url path - will overwrite existing values
+ *
+ * @param  {string} url
+ * @param  {string} qsName
+ * @param  {string} qsValue
+ * @return {string}
+ */
+utilities.appendQueryString = function (url, qsName, qsValue) {
+  const urlObj = Url.parse(url);
+  if (qsName && qsValue) {
+    urlObj.query = Querystring.parse(qsName + '=' + qsValue);
+    urlObj.search = '?' + encodeURIComponent(qsName) + '=' + encodeURIComponent(qsValue);
+  } else {
+    urlObj.search = '';
+  }
+
+  return urlObj.format(urlObj);
 };
